Validate GitHub auth as part of config
